feat: add not-found page and background images
