+++ src/components/ui/app-navbar.tsx
@@ -1,0 +1,83 @@
+"use client";
+
+import { Button } from "@nextui-org/button";
+import {
+  Navbar,
+  NavbarBrand,
+  NavbarContent,
+  NavbarItem,
+  NavbarMenu,
+  NavbarMenuItem,
+  NavbarMenuToggle,
+} from "@nextui-org/navbar";
+import { IconBook } from "@tabler/icons-react";
+
+import Link from "next/link";
+import React from "react";
+
+import { ThemeSwitcher } from "@/components/ui/theme-switcher";
+
+export default function AppNavbar() {
+  const [isMenuOpen, setIsMenuOpen] = React.useState(false);
+  const menuItems = [
+    {
+      label: "Home",
+      href: "/",
+    },
+    {
+      label: "Quiz",
+      href: "/quiz",
+    },
+  ];
+
+  return (
+    <div className="flex max-h-screen w-full flex-col">
+      <div></div>
+
+      <Navbar onMenuOpenChange={setIsMenuOpen}>
+        <NavbarContent>
+          <NavbarMenuToggle
+            aria-label={isMenuOpen ? "Close menu" : "Open menu"}
+            className="sm:hidden"
+          />
+          <NavbarBrand>
+            <IconBook size={28} />
+            <p className="p-2 font-bold text-inherit">QUIZ</p>
+          </NavbarBrand>
+        </NavbarContent>
+
+        <NavbarContent className="hidden gap-4 sm:flex" justify="center">
+          {menuItems.map((item, index) => (
+            <NavbarItem key={`${item}-${index}`}>
+              <Link className="w-full" href={item.href}>
+                {item.label}
+              </Link>
+            </NavbarItem>
+          ))}
+        </NavbarContent>
+        <NavbarContent justify="end">
+          <NavbarItem className="hidden lg:flex">
+            <Link href="#">Login</Link>
+          </NavbarItem>
+          <NavbarItem>
+            <Button as={Link} color="primary" href="#" variant="flat">
+              Sign Up
+            </Button>
+          </NavbarItem>
+          <NavbarItem>
+            <ThemeSwitcher />
+          </NavbarItem>
+        </NavbarContent>
+        <NavbarMenu>
+          {menuItems.map((item, index) => (
+            <NavbarMenuItem key={`${item}-${index}`}>
+              <Link className="w-full" href="{item.href}">
+                {item.label}
+              </Link>
+            </NavbarMenuItem>
+          ))}
+        </NavbarMenu>
+      </Navbar>
+    </div>
+  );
+}
